requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
pdfplumber>=0.9.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
langchain>=0.0.200
langchain-community>=0.0.10
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

